Consolidate CLI-related code into cli_tui.py, move folder security check into cli_commands.py and update import references
